+++ unicorn_binance_websocket_api/api/futures.py
@@ -40,10 +40,6 @@ class BinanceWebSocketApiApiFutures(object):
     stream restart, the payload is submitted as soon the stream is online again.
 
     Todo:
-        - https://binance-docs.github.io/apidocs/futures/en/#account-information-v2-user_data-2
-        - https://binance-docs.github.io/apidocs/futures/en/#account-information-user_data
-        - https://binance-docs.github.io/apidocs/futures/en/#futures-account-balance-v2-user_data-2
-        - https://binance-docs.github.io/apidocs/futures/en/#futures-account-balance-user_data
         - https://binance-docs.github.io/apidocs/futures/en/#position-information-v2-user_data-2
         - https://binance-docs.github.io/apidocs/futures/en/#position-information-user_data
         - https://binance-docs.github.io/apidocs/futures/en/#symbol-price-ticker-2
@@ -529,6 +525,421 @@ class BinanceWebSocketApiApiFutures(object):
             return new_client_order_id, response_value
 
         return new_client_order_id
+
+
+    def get_account_balance(self, process_response=None, recv_window: int = None, request_id: str = None,
+                           return_response: bool = False, stream_id: str = None, stream_label: str = None,
+                           version: Optional[Literal['v2']] = None) \
+            -> Union[str, dict, bool]:
+        """
+        Get your account balance.
+
+        Weight: 5
+
+        Official documentation:
+
+            - https://binance-docs.github.io/apidocs/futures/en/#futures-account-balance-user_data
+            - https://binance-docs.github.io/apidocs/futures/en/#futures-account-balance-v2-user_data-2
+
+        :param process_response: Provide a function/method to process the received webstream data (callback)
+                                 of this specific request.
+        :type process_response: function
+        :param recv_window: An additional parameter, `recvWindow`, may be sent to specify the number of milliseconds
+                            after timestamp the request is valid for. If `recvWindow` is not sent, it defaults to 5000.
+                            The value cannot be greater than 60000.
+        :type recv_window: int
+        :param request_id: Provide a custom id for the request
+        :type request_id: str
+        :param return_response: If `True` the response of the API request is waited for and returned directly.
+                                However, this increases the execution time of the function by the duration until the
+                                response is received from the Binance API.
+        :type return_response: bool
+        :param stream_id: ID of a stream to send the request
+        :type stream_id: str
+        :param stream_label: Label of a stream to send the request. Only used if `stream_id` is not provided!
+        :type stream_label: str
+        :param version: if None (default) method `account.status` is used, if 'v2' then `v2/account.status` is used.
+        :type version: str
+
+        :return: str, dict, bool
+
+        Message sent:
+
+        .. code-block:: json
+
+            {
+                "id": "605a6d20-6588-4cb9-afa0-b0ab087507ba",
+                "method": "account.balance",
+                "params": {
+                    "apiKey": "xTaDyrmvA9XT2oBHHjy39zyPzKCvMdtH3b9q4xadkAg2dNSJXQGCxzui26L823W2",
+                    "timestamp": 1702561978458,
+                    "signature": "208bb94a26f99aa122b1319490ca9cb2798fccc81d9b6449521a26268d53217a"
+                }
+            }
+
+        Response:
+
+        .. code-block:: json
+
+            {
+                "id": "605a6d20-6588-4cb9-afa0-b0ab087507ba",
+                "status": 200,
+                "result": {
+                    [
+                        {
+                            "accountAlias": "SgsR",    // unique account code
+                            "asset": "USDT",    // asset name
+                            "balance": "122607.35137903", // wallet balance
+                            "crossWalletBalance": "23.72469206", // crossed wallet balance
+                            "crossUnPnl": "0.00000000"  // unrealized profit of crossed positions
+                            "availableBalance": "23.72469206",       // available balance
+                            "maxWithdrawAmount": "23.72469206",     // maximum amount for transfer out
+                            "marginAvailable": true,    // whether the asset can be used as margin in Multi-Assets mode
+                            "updateTime": 1617939110373
+                        }
+                    ]
+                },
+                "rateLimits": [
+                  {
+                    "rateLimitType": "REQUEST_WEIGHT",
+                    "interval": "MINUTE",
+                    "intervalNum": 1,
+                    "limit": 2400,
+                    "count": 20
+                  }
+                ]
+            }
+
+        """
+        if stream_id is None:
+            if stream_label is not None:
+                stream_id = self._manager.get_stream_id_by_label(stream_label=stream_label)
+            else:
+                stream_id = self._manager.get_the_one_active_websocket_api()
+            if stream_id is None:
+                logger.critical(f"BinanceWebSocketApiApiFutures.get_account_balance() - error_msg: No `stream_id` "
+                                f"provided or found!")
+                return False
+
+        params = {"apiKey": self._manager.stream_list[stream_id]['api_key'],
+                  "timestamp": self._manager.get_timestamp()}
+
+        if recv_window is not None:
+            params['recvWindow'] = str(recv_window)
+
+        method = "account.balance" if version is None else f"{version}/account.balance"
+
+        api_secret = self._manager.stream_list[stream_id]['api_secret']
+        request_id = self._manager.get_new_uuid_id() if request_id is None else request_id
+        params['signature'] = self._manager.generate_signature(api_secret=api_secret, data=params)
+
+        payload = {"id": request_id,
+                   "method": method,
+                   "params": params}
+
+        if self._manager.send_with_stream(stream_id=stream_id, payload=payload) is False:
+            self._manager.add_payload_to_stream(stream_id=stream_id, payload=payload)
+
+        if process_response is not None:
+            with self._manager.process_response_lock:
+                entry = {'callback_function': process_response}
+                self._manager.process_response[request_id] = entry
+
+        if return_response is True:
+            with self._manager.return_response_lock:
+                entry = {'event_return_response': threading.Event()}
+                self._manager.return_response[request_id] = entry
+            self._manager.return_response[request_id]['event_return_response'].wait()
+            with self._manager.return_response_lock:
+                response_value = self._manager.return_response[request_id]['response_value']
+                del self._manager.return_response[request_id]
+            return response_value
+
+        return True
+
+    def get_account_status(self, process_response=None, recv_window: int = None, request_id: str = None,
+                           return_response: bool = False, stream_id: str = None, stream_label: str = None,
+                           version: Optional[Literal['v2']] = None) \
+            -> Union[str, dict, bool]:
+        """
+        Get current account information. User in single-asset/ multi-assets mode will see different value, see comments
+        in response section for detail.
+
+        Weight: 5
+
+        Official documentation:
+
+            - https://binance-docs.github.io/apidocs/futures/en/#account-information-user_data
+            - https://binance-docs.github.io/apidocs/futures/en/#account-information-v2-user_data-2
+
+        :param process_response: Provide a function/method to process the received webstream data (callback)
+                                 of this specific request.
+        :type process_response: function
+        :param recv_window: An additional parameter, `recvWindow`, may be sent to specify the number of milliseconds
+                            after timestamp the request is valid for. If `recvWindow` is not sent, it defaults to 5000.
+                            The value cannot be greater than 60000.
+        :type recv_window: int
+        :param request_id: Provide a custom id for the request
+        :type request_id: str
+        :param return_response: If `True` the response of the API request is waited for and returned directly.
+                                However, this increases the execution time of the function by the duration until the
+                                response is received from the Binance API.
+        :type return_response: bool
+        :param stream_id: ID of a stream to send the request
+        :type stream_id: str
+        :param stream_label: Label of a stream to send the request. Only used if `stream_id` is not provided!
+        :type stream_label: str
+        :param version: if None (default) method `account.status` is used, if 'v2' then `v2/account.status` is used.
+        :type version: str
+
+        :return: str, dict, bool
+
+        Message sent:
+
+        .. code-block:: json
+
+            {
+                "id": "605a6d20-6588-4cb9-afa0-b0ab087507ba",
+                "method": "account.status",
+                "params": {
+                    "apiKey": "xTaDyrmvA9XT2oBHHjy39zyPzKCvMdtH3b9q4xadkAg2dNSJXQGCxzui26L823W2",
+                    "timestamp": 1702620814781,
+                    "signature": "6bb98ef84170c70ba3d01f44261bfdf50fef374e551e590de22b5c3b729b1d8c"
+                }
+            }
+
+        Response Single Asset Mode:
+
+        .. code-block:: json
+
+            {
+              "id": "605a6d20-6588-4cb9-afa0-b0ab087507ba",
+              "status": 200,
+              "result": {
+                "feeTier": 0,       // account commission tier
+                "canTrade": true,   // if can trade
+                "canDeposit": true,     // if can transfer in asset
+                "canWithdraw": true,    // if can transfer out asset
+                "updateTime": 0,        // reserved property, please ignore
+                "multiAssetsMargin": false,
+                "tradeGroupId": -1,
+                "totalInitialMargin": "0.00000000",    // total initial margin required with current mark price (useless with isolated positions), only for USDT asset
+                "totalMaintMargin": "0.00000000",     // total maintenance margin required, only for USDT asset
+                "totalWalletBalance": "23.72469206",     // total wallet balance, only for USDT asset
+                "totalUnrealizedProfit": "0.00000000",   // total unrealized profit, only for USDT asset
+                "totalMarginBalance": "23.72469206",     // total margin balance, only for USDT asset
+                "totalPositionInitialMargin": "0.00000000",    // initial margin required for positions with current mark price, only for USDT asset
+                "totalOpenOrderInitialMargin": "0.00000000",   // initial margin required for open orders with current mark price, only for USDT asset
+                "totalCrossWalletBalance": "23.72469206",      // crossed wallet balance, only for USDT asset
+                "totalCrossUnPnl": "0.00000000",      // unrealized profit of crossed positions, only for USDT asset
+                "availableBalance": "23.72469206",       // available balance, only for USDT asset
+                "maxWithdrawAmount": "23.72469206"     // maximum amount for transfer out, only for USDT asset
+                "assets": [
+                    {
+                        "asset": "USDT",            // asset name
+                        "walletBalance": "23.72469206",      // wallet balance
+                        "unrealizedProfit": "0.00000000",    // unrealized profit
+                        "marginBalance": "23.72469206",      // margin balance
+                        "maintMargin": "0.00000000",        // maintenance margin required
+                        "initialMargin": "0.00000000",    // total initial margin required with current mark price
+                        "positionInitialMargin": "0.00000000",    //initial margin required for positions with current mark price
+                        "openOrderInitialMargin": "0.00000000",   // initial margin required for open orders with current mark price
+                        "crossWalletBalance": "23.72469206",      // crossed wallet balance
+                        "crossUnPnl": "0.00000000"       // unrealized profit of crossed positions
+                        "availableBalance": "23.72469206",       // available balance
+                        "maxWithdrawAmount": "23.72469206",     // maximum amount for transfer out
+                        "marginAvailable": true,    // whether the asset can be used as margin in Multi-Assets mode
+                        "updateTime": 1625474304765 // last update time
+                    },
+                    {
+                        "asset": "BUSD",            // asset name
+                        "walletBalance": "103.12345678",      // wallet balance
+                        "unrealizedProfit": "0.00000000",    // unrealized profit
+                        "marginBalance": "103.12345678",      // margin balance
+                        "maintMargin": "0.00000000",        // maintenance margin required
+                        "initialMargin": "0.00000000",    // total initial margin required with current mark price
+                        "positionInitialMargin": "0.00000000",    //initial margin required for positions with current mark price
+                        "openOrderInitialMargin": "0.00000000",   // initial margin required for open orders with current mark price
+                        "crossWalletBalance": "103.12345678",      // crossed wallet balance
+                        "crossUnPnl": "0.00000000"       // unrealized profit of crossed positions
+                        "availableBalance": "103.12345678",       // available balance
+                        "maxWithdrawAmount": "103.12345678",     // maximum amount for transfer out
+                        "marginAvailable": true,    // whether the asset can be used as margin in Multi-Assets mode
+                        "updateTime": 1625474304765 // last update time
+                    }
+                ],
+                "positions": [  // positions of all symbols in the market are returned
+                    // only "BOTH" positions will be returned with One-way mode
+                    // only "LONG" and "SHORT" positions will be returned with Hedge mode
+                    {
+                        "symbol": "BTCUSDT",    // symbol name
+                        "initialMargin": "0",   // initial margin required with current mark price
+                        "maintMargin": "0",     // maintenance margin required
+                        "unrealizedProfit": "0.00000000",  // unrealized profit
+                        "positionInitialMargin": "0",      // initial margin required for positions with current mark price
+                        "openOrderInitialMargin": "0",     // initial margin required for open orders with current mark price
+                        "leverage": "100",      // current initial leverage
+                        "isolated": true,       // if the position is isolated
+                        "entryPrice": "0.00000",    // average entry price
+                        "maxNotional": "250000",    // maximum available notional with current leverage
+                        "bidNotional": "0",  // bids notional, ignore
+                        "askNotional": "0",  // ask notional, ignore
+                        "positionSide": "BOTH",     // position side
+                        "positionAmt": "0",         // position amount
+                        "updateTime": 0           // last update time
+                    }
+                ]
+              },
+              "rateLimits": [
+                {
+                  "rateLimitType": "REQUEST_WEIGHT",
+                  "interval": "MINUTE",
+                  "intervalNum": 1,
+                  "limit": 2400,
+                  "count": 20
+                }
+              ]
+            }
+
+        Response Multi-Asset Mode:
+
+        .. code-block:: json
+
+            {
+              "id": "605a6d20-6588-4cb9-afa0-b0ab087507ba",
+              "status": 200,
+              "result": {
+                  "feeTier": 0,       // account commission tier
+                  "canTrade": true,   // if can trade
+                  "canDeposit": true,     // if can transfer in asset
+                  "canWithdraw": true,    // if can transfer out asset
+                  "updateTime": 0,        // reserved property, please ignore
+                  "multiAssetsMargin": true,
+                  "tradeGroupId": -1,
+                  "totalInitialMargin": "0.00000000",    // the sum of USD value of all cross positions/open order initial margin
+                  "totalMaintMargin": "0.00000000",     // the sum of USD value of all cross positions maintenance margin
+                  "totalWalletBalance": "126.72469206",     // total wallet balance in USD
+                  "totalUnrealizedProfit": "0.00000000",   // total unrealized profit in USD
+                  "totalMarginBalance": "126.72469206",     // total margin balance in USD
+                  "totalPositionInitialMargin": "0.00000000",    // the sum of USD value of all cross positions initial margin
+                  "totalOpenOrderInitialMargin": "0.00000000",   // initial margin required for open orders with current mark price in USD
+                  "totalCrossWalletBalance": "126.72469206",      // crossed wallet balance in USD
+                  "totalCrossUnPnl": "0.00000000",      // unrealized profit of crossed positions in USD
+                  "availableBalance": "126.72469206",       // available balance in USD
+                  "maxWithdrawAmount": "126.72469206"     // maximum virtual amount for transfer out in USD
+                  "assets": [
+                      {
+                          "asset": "USDT",            // asset name
+                          "walletBalance": "23.72469206",      // wallet balance
+                          "unrealizedProfit": "0.00000000",    // unrealized profit
+                          "marginBalance": "23.72469206",      // margin balance
+                          "maintMargin": "0.00000000",        // maintenance margin required
+                          "initialMargin": "0.00000000",    // total initial margin required with current mark price
+                          "positionInitialMargin": "0.00000000",    //initial margin required for positions with current mark price
+                          "openOrderInitialMargin": "0.00000000",   // initial margin required for open orders with current mark price
+                          "crossWalletBalance": "23.72469206",      // crossed wallet balance
+                          "crossUnPnl": "0.00000000"       // unrealized profit of crossed positions
+                          "availableBalance": "126.72469206",       // available balance
+                          "maxWithdrawAmount": "23.72469206",     // maximum amount for transfer out
+                          "marginAvailable": true,    // whether the asset can be used as margin in Multi-Assets mode
+                          "updateTime": 1625474304765 // last update time
+                      },
+                      {
+                          "asset": "BUSD",            // asset name
+                          "walletBalance": "103.12345678",      // wallet balance
+                          "unrealizedProfit": "0.00000000",    // unrealized profit
+                          "marginBalance": "103.12345678",      // margin balance
+                          "maintMargin": "0.00000000",        // maintenance margin required
+                          "initialMargin": "0.00000000",    // total initial margin required with current mark price
+                          "positionInitialMargin": "0.00000000",    //initial margin required for positions with current mark price
+                          "openOrderInitialMargin": "0.00000000",   // initial margin required for open orders with current mark price
+                          "crossWalletBalance": "103.12345678",      // crossed wallet balance
+                          "crossUnPnl": "0.00000000"       // unrealized profit of crossed positions
+                          "availableBalance": "126.72469206",       // available balance
+                          "maxWithdrawAmount": "103.12345678",     // maximum amount for transfer out
+                          "marginAvailable": true,    // whether the asset can be used as margin in Multi-Assets mode
+                          "updateTime": 1625474304765 // last update time
+                      }
+                  ],
+                  "positions": [  // positions of all symbols in the market are returned
+                      // only "BOTH" positions will be returned with One-way mode
+                      // only "LONG" and "SHORT" positions will be returned with Hedge mode
+                      {
+                          "symbol": "BTCUSDT",    // symbol name
+                          "initialMargin": "0",   // initial margin required with current mark price
+                          "maintMargin": "0",     // maintenance margin required
+                          "unrealizedProfit": "0.00000000",  // unrealized profit
+                          "positionInitialMargin": "0",      // initial margin required for positions with current mark price
+                          "openOrderInitialMargin": "0",     // initial margin required for open orders with current mark price
+                          "leverage": "100",      // current initial leverage
+                          "isolated": true,       // if the position is isolated
+                          "entryPrice": "0.00000",    // average entry price
+                          "breakEvenPrice": "0.0",    // average entry price
+                          "maxNotional": "250000",    // maximum available notional with current leverage
+                          "bidNotional": "0",  // bids notional, ignore
+                          "askNotional": "0",  // ask notional, ignore
+                          "positionSide": "BOTH",     // position side
+                          "positionAmt": "0",         // position amount
+                          "updateTime": 0           // last update time
+                      }
+                  ]
+              },
+              "rateLimits": [
+                {
+                  "rateLimitType": "REQUEST_WEIGHT",
+                  "interval": "MINUTE",
+                  "intervalNum": 1,
+                  "limit": 2400,
+                  "count": 20
+                }
+              ]
+            }
+        """
+        if stream_id is None:
+            if stream_label is not None:
+                stream_id = self._manager.get_stream_id_by_label(stream_label=stream_label)
+            else:
+                stream_id = self._manager.get_the_one_active_websocket_api()
+            if stream_id is None:
+                logger.critical(f"BinanceWebSocketApiApiFutures.get_account_status() - error_msg: No `stream_id` "
+                                f"provided or found!")
+                return False
+
+        params = {"apiKey": self._manager.stream_list[stream_id]['api_key'],
+                  "timestamp": self._manager.get_timestamp()}
+
+        if recv_window is not None:
+            params['recvWindow'] = str(recv_window)
+
+        method = "account.status" if version is None else f"{version}/account.status"
+
+        api_secret = self._manager.stream_list[stream_id]['api_secret']
+        request_id = self._manager.get_new_uuid_id() if request_id is None else request_id
+        params['signature'] = self._manager.generate_signature(api_secret=api_secret, data=params)
+
+        payload = {"id": request_id,
+                   "method": method,
+                   "params": params}
+
+        if self._manager.send_with_stream(stream_id=stream_id, payload=payload) is False:
+            self._manager.add_payload_to_stream(stream_id=stream_id, payload=payload)
+
+        if process_response is not None:
+            with self._manager.process_response_lock:
+                entry = {'callback_function': process_response}
+                self._manager.process_response[request_id] = entry
+
+        if return_response is True:
+            with self._manager.return_response_lock:
+                entry = {'event_return_response': threading.Event()}
+                self._manager.return_response[request_id] = entry
+            self._manager.return_response[request_id]['event_return_response'].wait()
+            with self._manager.return_response_lock:
+                response_value = self._manager.return_response[request_id]['response_value']
+                del self._manager.return_response[request_id]
+            return response_value
+
+        return True
 
     def get_order(self, order_id: int = None, orig_client_order_id: str = None, process_response=None,
                   recv_window: int = None, request_id: str = None, return_response: bool = False, stream_id: str = None,