+++ dev/sphinx/source/changelog.md
@@ -21,8 +21,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
   - `manager.api.spot.get_recent_trades()`
   - `manager.api.spot.get_unfilled_order_count()`
 - Support for `Websocket API Futures`:
+  - `manager.api.futures.get_account_status()`
   - `manager.api.futures.cancel_order()`
   - `manager.api.futures.create_order()`
+  - `manager.api.futures.get_account_status()` (v1+v2)
+  - `manager.api.futures.get_account_balance()` (v1+v2)
   - `manager.api.futures.get_order()` 
   - `manager.api.futures.get_order_book()`
   - `manager.api.futures.modify_order()`